summarize-monologue/requirements.txt
functions-framework==3.4.0
google-cloud-storage==2.13.0
google-cloud-firestore==2.13.0
google-genai==1.5.0
requests==2.31.0
pydantic==2.5.2